Prepare for TOMATO SSD model
* Handle more than 256 nodes
* Drop the check for the number of inputs
* Dynamically add operators from the model
* Dynamic shape inference
    * Make models compatible - fix hard-coded Reshape in mnist-8 and
      remove audio preprocessing steps from KWS_DNN
    * Always use 4-D tensors as input data instead of an array of 4-D
      tensors
* Much faster way to get intermediate OFMs from onnxruntime - needed
  for models with hunders of layers like SSD
    * Remove the no longer needed script for parsing onnxruntime
      debugging outputs
* Handle nodes with multiple outputs
* Dynamically determine the maximum number of node inputs
* determine_conv_tile_c: Handle incomplete channel tilling
* parse_model improvements
    * Handle ValueInfoProto messages without TypeProto
    * Print memory usage for each OFM
    * Print contents of INT64 tensors
    * Accept even larger models

+++ common/cnn_common.cpp
@@ -5,7 +5,7 @@
 
 ParameterInfo intermediate_parameters_info_vm[MODEL_NODES_LEN];
 
-const ParameterInfo* get_parameter_info(uint8_t i) {
+const ParameterInfo* get_parameter_info(uint16_t i) {
     if (i < N_INPUT) {
         return reinterpret_cast<const ParameterInfo*>(model_parameters_info_data) + i;
     } else {
